usmap: US maps including Alaska and Hawaii

It is usually difficult or inconvenient to create US maps that include both Alaska and Hawaii in a convenient spot. All map data frames produced by this package use the Albers Equal Area projection.

Map data frames

Alaska and Hawaii have been manually moved to a new location so that their new coordinates place them to the bottom-left corner of the map. These maps can be accessed by using the us_map function.

The function provides the ability to retrieve maps with either state borders or county borders using the regions parameter for convenience.

States (or counties) can be included and excluded using the provided include and exclude parameters. These parameters can be used together with any combination of names, abbreviations, or FIPS code to create more complex maps.

FIPS lookup tools

Several functions have been included to lookup the US state or county pertaining to a FIPS code.

Likewise a reverse lookup can be done where a FIPS code can be used to retrieve the associated state(s) or county(ies). This can be useful when preparing data to be merged with the map data frame.

Plot US map data

A convenience function plot_usmap has been included which takes similar parameters to us_map and returns a ggplot2::ggplot2 object. Since the output is a ggplot object, other layers can be added such as scales, themes, and labels. Including data in the function call will color the map according to the values in the data, creating a choropleth.

References

Rudis, Bob. "Moving The Earth (well, Alaska & Hawaii) With R." Blog post. Rud.is., 16 Nov. 2014. Web. 10 Aug. 2015. https://rud.is/b/2014/11/16/moving-the-earth-well-alaska-hawaii-with-r/.

See also

Author

Paolo Di Lorenzo

Email: dilorenzo@hey

GitHub: https://github.com/pdil/