Convert coordinate date frame to usmap projection

Converting an external data frame of map coordinates will allow those points to line up with the regular usmap plot by applying the same Albers Equal Area projection to those points as well.

Usage

usmap_transform(
  data,
  input_names = c("lon", "lat"),
  output_names = c("x", "y")
)

# S3 method for data.frame
usmap_transform(
  data,
  input_names = c("lon", "lat"),
  output_names = c("x", "y")
)

Arguments

data

A data frame containing coordinates in a two column format where the first column represents longitude and the second data frame represents latitude. The names of the data frame column do not matter, just that the order of the columns is kept intact.

input_names

A character vector of length two which specifies the longitude and latitude columns of the input data (the ones that should be transformed), respectively. Defaults to c("lon", "lat").

output_names

A character vector of length two which specifies the longitude and latitude columns of the output data (after transformation), respectively. Defaults to c("x", "y").

Value

A data frame containing the transformed coordinates from the input data frame with the Albers Equal Area projection applied. The transformed columns will be appended to the data frame so that all original columns should remain intact.

Examples

data <- data.frame(
  lon = c(-74.01, -95.36, -118.24, -87.65, -134.42, -157.86),
  lat = c(40.71, 29.76, 34.05, 41.85, 58.30, 21.31),
  pop = c(8398748, 2325502, 3990456, 2705994, 32113, 347397)
)

# Transform data
transformed_data <- usmap_transform(data)
#> Warning: Evolution status fixed to 2L
#> Error in usmap_transform(data): `sf` must be installed to use `usmap_transform`.
#>          Use: install.packages("sf") and try again.

# Plot transformed data on map
library(ggplot2)

plot_usmap() + geom_point(
  data = transformed_data,
  aes(x = x, y = y, size = pop),
  color = "red", alpha = 0.5
)
#> Error in fortify(data): object 'transformed_data' not found